requirements.txt
Abg==2.3.1
deep-translator
pyromod
openai
aiofiles
aiohttp
asyncio
beautifulsoup4
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
httpx==0.27.2
lyricsgenius
motor
wget
pillow==9.5.0
psutil
pytube
pykeyboard
unidecode
py-tgcalls==1.2.9
ntgcalls==1.1.2
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
tgcrypto
urllib3
uvloop
git+https://github.com/KurimuzonAkuma/pyrogram@dev
git+https://github.com/yt-dlp/yt-dlp
youtube-search-python
git+https://github.com/joetats/youtube_search@master
git+https://github.com/TheChampu/yt-dlp-youtube-oauth2